添加生成动态区域逻辑

+++ src/main/java/com/astra/actionconfig/config/data/DynamicArea.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 @Data
 public class DynamicArea {
     public String id;
-    public int heightToWidthRatio = 1;
+    public double heightToWidthRatio = 1;
     public float width = 0.1f;
     public Point2F[] area = new Point2F[4];
     public double[] limitedArea = new double[4];
